commit quatre

+++ pages/supprimerFiliere.php
@@ -1,22 +1,25 @@
-<?php require_once "connexiondb.php"; 
+<?php
+session_start();
+if (isset($_SESSION['utilisateur'])) {
+    require_once "connexiondb.php";
 
-$idf = isset($_GET['idF'])?$_GET['idF']:0;
+    $idf = isset($_GET['idF']) ? $_GET['idF'] : 0;
 
-$requeteStag = "SELECT COUNT(*) countStag FROM stagiaire WHERE idFiliere=$idf";
-$resultatStag = $pdo->query($requeteStag);
-$tabCountStag = $resultatStag->fetch();
-$nbrStag = $tabCountStag['countStag'];
+    $requeteStag = "SELECT COUNT(*) countStag FROM stagiaire WHERE idFiliere=$idf";
+    $resultatStag = $pdo->query($requeteStag);
+    $tabCountStag = $resultatStag->fetch();
+    $nbrStag = $tabCountStag['countStag'];
 
-if ($nbrStag == 0) {
-    $requete = "DELETE FROM filiere WHERE idFiliere=?";
-    $params = array($idf);
-    $resultat = $pdo->prepare($requete);
-    $resultat->execute($params);
-    header("location: filieres.php");
-}else {
-    $msg = "Suppression impossible : Vous devez d'abord supprimer tous les stagiaires incris dans cette filière";
-    header("location: alerte.php?message=$msg");
+    if ($nbrStag == 0) {
+        $requete = "DELETE FROM filiere WHERE idFiliere=?";
+        $params = array($idf);
+        $resultat = $pdo->prepare($requete);
+        $resultat->execute($params);
+        header("location: filieres.php");
+    } else {
+        $msg = "Suppression impossible : Vous devez d'abord supprimer tous les stagiaires incris dans cette filière";
+        header("location: alerte.php?message=$msg");
+    }
+} else {
+    header("location: login.php");
 }
-
-?>
-
